% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadataconfig.R
\name{aldatata.api_key}
\alias{aldatata.api_key}
\title{Query or set Altadata API key}
\usage{
aldatata.api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Altadata \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set Altadata API key
}
\examples{
\dontrun{
aldatata.api_key('foobar')
}
}
