% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.condition_in}
\alias{altadata.condition_in}
\title{In condition by given column and value list}
\usage{
altadata.condition_in(condition_column, condition_value)
}
\arguments{
\item{condition_column}{column to which the condition will be applied}

\item{condition_value}{value to use with condition}
}
\value{
Nothing just set the in condition parameters
}
\description{
In condition by given column and value list
}
\examples{
\dontrun{
  altadata.condition_in("province_state", c("Utah", "Alabama"))
}
}
