% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.equal}
\alias{altadata.equal}
\title{Equal condition by given column and value in the retrieve data process}
\usage{
altadata.equal(condition_column, condition_value)
}
\arguments{
\item{condition_column}{column to which the condition will be applied}

\item{condition_value}{value to use with condition}
}
\value{
Nothing just set the equal condition parameters
}
\description{
Equal condition by given column and value in the retrieve data process
}
\examples{
\dontrun{
  altadata.equal("province_state", "Alabama")
}
}
