% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.get_data}
\alias{altadata.get_data}
\title{Initialize retrieve data process}
\usage{
altadata.get_data(product_code, limit)
}
\arguments{
\item{product_code}{data product code}

\item{limit}{number of rows you want to retrieve}
}
\value{
Nothing just set the initial parameters
}
\description{
Initialize retrieve data process
}
\examples{
\dontrun{
  altadata.get_data("co_10_jhucs_03", limit = 50)
}
}
