% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.get_header}
\alias{altadata.get_header}
\title{Get data header as a vector}
\usage{
altadata.get_header(product_code)
}
\arguments{
\item{product_code}{data product code}
}
\value{
vector object
}
\description{
Get data header as a vector
}
\examples{
\dontrun{
  aldatata.api_key('YOUR_API_KEY')
  altadata.get_header("co_10_jhucs_03")
}
}
