% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.greater_than}
\alias{altadata.greater_than}
\title{Greater than condition by given column and value}
\usage{
altadata.greater_than(condition_column, condition_value)
}
\arguments{
\item{condition_column}{column to which the condition will be applied}

\item{condition_value}{value to use with condition}
}
\value{
Nothing just set the greater than condition parameters
}
\description{
Greater than condition by given column and value
}
\examples{
\dontrun{
  altadata.greater_than("mortality_rate", 2)
}
}
