% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.less_than}
\alias{altadata.less_than}
\title{Less than condition by given column and value}
\usage{
altadata.less_than(condition_column, condition_value)
}
\arguments{
\item{condition_column}{column to which the condition will be applied}

\item{condition_value}{value to use with condition}
}
\value{
Nothing just set the less than condition parameters
}
\description{
Less than condition by given column and value
}
\examples{
\dontrun{
  altadata.less_than("mortality_rate", 2)
}
}
