% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.list_subscription}
\alias{altadata.list_subscription}
\title{Get customer subscription info}
\usage{
altadata.list_subscription()
}
\value{
dataframe object
}
\description{
Get customer subscription info
}
\examples{
\dontrun{
  altadata.list_subscription()
}
}
