% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.load}
\alias{altadata.load}
\title{Fetch data with configurations given before}
\usage{
altadata.load()
}
\value{
dataframe object
}
\description{
Fetch data with configurations given before
}
\examples{
\dontrun{
  aldatata.api_key('YOUR_API_KEY')
  altadata.get_data("co_10_jhucs_03", limit = 50)
  altadata.load()
}
}
