% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.not_equal}
\alias{altadata.not_equal}
\title{Not equal condition by given column and value}
\usage{
altadata.not_equal(condition_column, condition_value)
}
\arguments{
\item{condition_column}{column to which the condition will be applied}

\item{condition_value}{value to use with condition}
}
\value{
Nothing just set the not equal condition parameters
}
\description{
Not equal condition by given column and value
}
\examples{
\dontrun{
  altadata.not_equal("province_state", "Utah")
}
}
