% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altadata.R
\name{altadata.select}
\alias{altadata.select}
\title{Select specific columns in the retrieve data process}
\usage{
altadata.select(selected_columns)
}
\arguments{
\item{selected_columns}{list of columns to select}
}
\value{
Nothing just set the select parameters
}
\description{
Select specific columns in the retrieve data process
}
\examples{
\dontrun{
  altadata.select(c("reported_date", "province_state", "mortality_rate"))
}
}
